+++ src/core/ontologyIO/OntologyLoader.cpp
@@ -186,7 +186,7 @@ namespace ontologenius {
   std::string OntologyLoader::downlaodFileCurl(const std::string& uri)
   {
     std::string res;
-    res.reserve(1024 * 1024);
+    res.reserve((size_t)1024 * (size_t)1024);
     curl_global_init(CURL_GLOBAL_ALL);
     CURL* curl_handle = curl_easy_init();
     curl_easy_setopt(curl_handle, CURLOPT_URL, uri.c_str());
