+++ src/core/reasoner/plugins/ReasonerGeneralize.cpp
@@ -5,6 +5,7 @@
 #include <pluginlib/class_list_macros.hpp>
 #include <shared_mutex>
 #include <string>
+#include <tuple>
 #include <unordered_set>
 #include <vector>
 
@@ -48,7 +49,7 @@ namespace ontologenius {
         std::shared_lock<std::shared_timed_mutex> lock_indiv_shared(ontology_->individual_graph_.mutex_);
         std::shared_lock<std::shared_timed_mutex> lock_shared(ontology_->class_graph_.mutex_);
 
-        PropertiesCounter<DataPropertyBranch*, LiteralNode*> data_counter(min_count_, min_percent_);
+        PropertiesCounter<DataPropertyBranch*, LiteralNode*> data_counter((int)min_count_, min_percent_);
         PropertiesCounter<ObjectPropertyBranch*, ClassBranch*> object_counter;
 
         for(auto* down : down_set)
@@ -111,7 +112,7 @@ namespace ontologenius {
               notifications_.emplace_back(notification_info, "[CHANGE]" + me->value() + ">" + std::get<0>(property)->value() + ":" + std::get<1>(property)->value());
 
             me->data_relations_[prop_i].second = std::get<1>(property);
-            me->data_relations_[prop_i].probability = (float)std::get<2>(property) - 0.01;
+            me->data_relations_[prop_i].probability = (float)std::get<2>(property) - 0.01f;
           }
         }
 
@@ -129,7 +130,7 @@ namespace ontologenius {
       for(auto i : deduced_indexs)
       {
         notifications_.emplace_back(notification_info, "[DELETE]" + me->value() + ">" + me->data_relations_[i - deleted].first->value() + ":" + me->data_relations_[i - deleted].second->value());
-        me->data_relations_.erase(me->data_relations_.begin() + i - deleted);
+        me->data_relations_.erase(me->data_relations_.begin() + (int)i - (int)deleted);
         deleted++;
       }
     }
@@ -149,7 +150,7 @@ namespace ontologenius {
       for(size_t prop_i = 0; prop_i < me->object_relations_.size(); prop_i++)
         if(me->object_relations_[prop_i].first == std::get<0>(property))
         {
-          index = prop_i;
+          index = (int)prop_i;
           deduced_indexs.erase(index);
           if(me->object_relations_[prop_i].probability < 1.0)
           {
@@ -157,7 +158,7 @@ namespace ontologenius {
               notifications_.emplace_back(notification_info, "[CHANGE]" + me->value() + ">" + std::get<0>(property)->value() + ":" + std::get<1>(property)->value());
 
             me->object_relations_[prop_i].second = std::get<1>(property);
-            me->object_relations_[prop_i].probability = (float)std::get<2>(property) - 0.01;
+            me->object_relations_[prop_i].probability = (float)std::get<2>(property) - 0.01f;
           }
         }
 
@@ -174,7 +175,7 @@ namespace ontologenius {
       for(auto i : deduced_indexs)
       {
         notifications_.emplace_back(notification_info, "[DELETE]" + me->value() + ">" + me->object_relations_[i - deleted].first->value() + ":" + me->object_relations_[i - deleted].second->value());
-        me->object_relations_.erase(me->object_relations_.begin() + i - deleted);
+        me->object_relations_.erase(me->object_relations_.begin() + (int)i - (int)deleted);
         deleted++;
       }
     }
