+++ src/graphical/ontoloGUI/ontologui.cpp
@@ -1,8 +1,5 @@
 #include "ontologenius/graphical/ontoloGUI/ontologui.h"
 
-#include <QMainWindow>
-#include <QObject>
-#include <QScrollBar>
 #include <algorithm>
 #include <cstddef>
 #include <regex>
@@ -13,8 +10,13 @@
 #include "ontologenius/graphical/ontoloGUI/qpushbuttonextended.h"
 #include "ui_ontologui.h"
 // Qt includes
+#include "qcheckbox.h"
 #include "qcoreevent.h"
+#include "qlayoutitem.h"
+#include "qmainwindow.h"
+#include "qobject.h"
 #include "qobjectdefs.h"
+#include "qscrollbar.h"
 #include "qstring.h"
 #include "qwidget.h"
 
@@ -834,7 +836,7 @@ void OntoloGUI::ontologyNameAddDelChangedSlot(const QString& text)
   {
     const size_t equal_pose = text.toStdString().find('=');
     if(equal_pose != std::string::npos)
-      ui_->OntologyName->setText(text.mid(0, equal_pose));
+      ui_->OntologyName->setText(text.mid(0, (int)equal_pose));
     else
       ui_->OntologyName->setText(text);
   }
